--- a/Module.mk
+++ b/Module.mk
@@ -104,6 +104,7 @@ include src/main/d3d9exhook/Module.mk
 include src/main/ddrhook-util/Module.mk
 include src/main/ddrhook1/Module.mk
 include src/main/ddrhook2/Module.mk
+include src/main/ddrio-async/Module.mk
 include src/main/ddrio-p3io/Module.mk
 include src/main/ddrio-mm/Module.mk
 include src/main/ddrio-smx/Module.mk
@@ -730,6 +731,7 @@ $(zipdir)/ddr-16-x64.zip: \
 	$(V)zip -j $@ $^
 
 $(zipdir)/ddr-hwio-x86.zip: \
+		build/bin/indep-32/ddrio-async.dll \
 		build/bin/indep-32/ddrio-p3io.dll \
 		build/bin/indep-32/ddrio-mm.dll \
 		build/bin/indep-32/ddrio-smx.dll \
@@ -741,6 +743,7 @@ $(zipdir)/ddr-hwio-x86.zip: \
 	$(V)zip -j $@ $^
 
 $(zipdir)/ddr-hwio-x64.zip: \
+		build/bin/indep-64/ddrio-async.dll \
 		build/bin/indep-64/ddrio-p3io.dll \
 		build/bin/indep-64/ddrio-mm.dll \
 		build/bin/indep-64/ddrio-smx.dll \