--- a/doc/ddrhook/README.md
+++ b/doc/ddrhook/README.md
@@ -46,7 +46,10 @@ IO hardware you want to use:
 * `ddrio`: Default implementation supporting keyboard, mouse and USB
 game controllers
 * ddrio-mm: Support Minimaid custom interface
-* [ddrio-smx](ddrhook/ddrio-smx.md): Support for StepManiaX dance platforms
+* [ddrio-smx](ddrio-smx.md): Support for StepManiaX dance platforms
+* [ddrio-p3io](ddrio-p3io.md): P3IO + EXTIO driver implementation
+* [ddrio-async](ddrio-async.md): Wrapper/shim library to drive another ddrio in
+  a dedicated IO thread
 
 ## Unicorntail