--- a/doc/release-process.md
+++ b/doc/release-process.md
@@ -1,18 +1,24 @@
 # Release process
 
-For bemanitools maintainers, steps for the release process:
+For maintainers, steps for the release process:
 
 1. Update the corresponding section in the [changelog](CHANGELOG.md) with bullet points reflecting
-major features merged into master since the previous release
+   major features merged into master since the previous release
 1. Make sure you have the latest `master` state pulled to your local copy
 1. Tag the release with the next release number which also be found at the
-[top of the readme](#bemanitools-5) as it always reflects the current version beinged worked on
-on the `master` branch , e.g. 5.35: `git tag 5.35`
+   [top of the readme](../README.md) as it always reflects the current version being worked on
+   on the `master` branch , e.g. 0.04: `git tag 5.35`
 1. Push the tag to upstream: `git push origin 5.35`
-1. The [build pipeline](https://dev.s-ul.net/djhackers/bemanitools/-/pipelines) should start
-automatically once the tag is pushed including the steps `build` and `upload-release`
-1. Once completed successfully, the release is uploaded
-1. Take the changelog of the published version and notify the pigs in the stall about it:
+1. The [build pipeline](https://github.com/djhackersdev/bemanitools/actions) should start
+   automatically once the tag is pushed including the steps `build` and `publish-release`
+1. Once completed successfully, the release is published on the
+   [releases page](https://github.com/djhackersdev/bemanitools/releases/)
+1. Edit the latest release
+1. Copy-pate the change log section of the respective version into the description
+1. Un-check *Set as pre-release*
+1. Check *Set as the latest release*
+1. Click *Update release*
+1. Notify any other channels, e.g. the pigs in the stall, about the latest release:
     1. New post in thread
         ```
         <insert version here> released: <direct link to published version>
@@ -23,7 +29,7 @@ automatically once the tag is pushed including the steps `build` and `upload-rel
     1. Update the OP title by bumping the version number in it
     1. Update the OP post by extending it accordingly (see previous entries)
 
-1. Bump the version number at the [top of this readme](#bemanitools-5) and add a new empty section
-in the [changelog](CHANGELOG.md) with the new version number as title. Commit changes and push to
+1. Bump the version number at the [top of the readme](../README.md) and add a new empty section
+in the [changelog](../CHANGELOG.md) with the new version number as title. Commit changes and push to
 `master` branch
-1. Continue developing and merging MRs until you decide its time for another release
+1. Continue developing and merging PRs until you decide its time for another release