--- a/src/main/ddrio-async/Module.mk
+++ b/src/main/ddrio-async/Module.mk
@@ -1,0 +1,9 @@
+dlls		+= ddrio-async
+
+ldflags_ddrio-async:= \
+
+libs_ddrio-async	:= \
+	util \
+
+src_ddrio-async	:= \
+	ddrio.c \