--- a/src/main/ddrio-async/ddrio-async.def
+++ b/src/main/ddrio-async/ddrio-async.def
@@ -1,0 +1,11 @@
+LIBRARY ddrio-async
+
+EXPORTS
+    ddr_io_set_loggers
+    ddr_io_fini
+    ddr_io_init
+    ddr_io_read_pad
+    ddr_io_set_lights_extio
+    ddr_io_set_lights_p3io
+    ddr_io_set_lights_hdxs_panel
+    ddr_io_set_lights_hdxs_rgb